+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Enhanced editor, now possible to use : bold, italics, underscore, quote, order/unordered list, link and embed image (all customizable), #47
 * Comment read more now considers content height instead of length as it includes HTML and website can be wide, which breaks logic of content length, #105
 * Added "*" for name and email fields to point that they are important to enter
+* Send comment button is toggled as disabled when agreement is not accepted (for guest and authorized users)
 
 **Fixes:**